move static_assert definitions in detail/asserts folder
clean zip_with a bit
