+++ include/fplus/pairs.hpp
@@ -9,33 +9,12 @@
 #include <fplus/container_common.hpp>
 #include <fplus/function_traits.hpp>
 #include <fplus/detail/invoke.hpp>
-#include <fplus/detail/function_traits_asserts.hpp>
+#include <fplus/detail/asserts/pairs.hpp>
 
 #include <utility>
 
 namespace fplus
 {
-namespace detail
-{
-struct zip_with_tag
-{
-};
-
-template <typename F, typename X, typename Y>
-struct function_traits_asserts<zip_with_tag, F, X, Y>
-{
-    static_assert(utils::function_traits<F>::arity == 2,
-        "Function must take two parameters.");
-    using FOut = invoke_result_t<F, X, Y>;
-    typedef typename utils::function_traits<F>::template arg<0>::type FIn0;
-    typedef typename utils::function_traits<F>::template arg<1>::type FIn1;
-    static_assert(std::is_convertible<X, FIn0>::value,
-        "Function does not take elements from first Container as first Parameter.");
-    static_assert(std::is_convertible<Y, FIn1>::value,
-        "Function does not take elements from second Container as second Parameter.");
-};
-}
-
 // API search type: apply_to_pair : (((a, b) -> c), (a, b)) -> c
 // fwd bind count: 1
 // Apply binary function to parts of a pair.
@@ -63,10 +42,6 @@ template <typename ContainerIn1,
           typename ContainerOut = std::vector<TOut>>
 ContainerOut zip_with(F f, const ContainerIn1& xs, const ContainerIn2& ys)
 {
-  using FOut = detail::invoke_result_t<F, X, Y>;
-  static_assert(
-      std::is_convertible<FOut, TOut>::value,
-      "Elements produced by this function can not be stored in ContainerOut.");
   static_assert(
       std::is_same<
           typename internal::same_cont_new_t<ContainerIn1, void>::type,
@@ -80,7 +55,7 @@ ContainerOut zip_with(F f, const ContainerIn1& xs, const ContainerIn2& ys)
   auto itYs = std::begin(ys);
   for (std::size_t i = 0; i < resultSize; ++i)
   {
-    *itResult = f(*itXs, *itYs);
+    *itResult = detail::invoke(f, *itXs, *itYs);
     ++itXs;
     ++itYs;
   }
